% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bias_absences.R
\name{create_bias_absences}
\alias{create_bias_absences}
\title{Function to create absences that account for sampling bias
Create biased pseudo-absence points.}
\usage{
create_bias_absences(species_data, raster_mask, process_boundary, abs_number,
  return_spatial = FALSE)
}
\arguments{
\item{species_data}{SpatialPointsDataFrame with species data.}

\item{raster_mask}{Raster mask with same processing resolution.}

\item{process_boundary}{SpatialPolygonsDataFrame with the processing extent.}

\item{abs_number}{Number of created pseudo-absences.}
}
\description{
Function to create absences that account for sampling bias
Create biased pseudo-absence points.
}
