% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarify_points.R
\name{rarify_points}
\alias{rarify_points}
\title{Rarify points
Rarify points according to the raster grid mask. Points will be rarified so only one 
point per grid cell is preserved. Purposfully made for rarifying points in presence-absence modeling. Useful for SDM.}
\usage{
rarify_points(point_data, raster_data)
}
\arguments{
\item{point_data}{Input point data to be rarified. SpatialPointsDataFrame.}

\item{raster_data}{Input grid data to which point_data is rarified. RasterLayer. Has to have overlapping extent.}
}
\value{
SpatialPointsDataFrame with rarified points.
}
\description{
Rarify points
Rarify points according to the raster grid mask. Points will be rarified so only one 
point per grid cell is preserved. Purposfully made for rarifying points in presence-absence modeling. Useful for SDM.
}
